Redoing the refactor, introduced bugs

+++ include/raytracer.h
@@ -16,7 +16,7 @@ class RayTracer {
   // Screen width and height in pixels
   size_t screen_width = 320, screen_height = 200;
   // Horizontal field of view, in degrees
-  double horizontal_fov = 60;
+  float horizontal_fov = 60;
   // Maximum recursive depth, for reflection ray-casts
   size_t max_depth = 10;
   // We super-sample each pixel in a square with aa_num points and average them
